requirements_1.txt
aiohttp==3.9.5
multiprocess==0.70.16
pandas==2.2.2
pdfservices-sdk==2.3.1
pillow==10.3.0
PyMuPDF==1.24.5
python-dotenv==1.0.1
streamsync==0.5.0